helm-chart: set SECRET_KEY from values file
I'm honestly not sure why this isn't working as-is (without this
change). The Django settings.py is reading the SECRET_KEY value from an
environment variable DJANGO_SECRET_KEY. Which is set. However, on
deployment, it says it is not.

Try setting SECRET_KEY itself. Maybe Django has some sort of built-in
"environment variable takes precedence over config file" override that
I'm unaware of?