Fix NGINX resource requests/limits
The NGINX component was added as part of "bolting on" S3 storage to this
project. The commit which added the NGINX component contained a typo: it
accidentally halved the memory limits configured for the MOP
application.

Fix the typo and restore the MOP application resource requests/limits to
their original level.