Automatically detect CPU count
Run each fit using all of the CPUs in the machine by default. The user
can override this value to be anything they want, to control the
parallelism without needing to edit the code to change this value.

This may be used in the future in the Kubernetes Helm charts to tune
resource usage across the cluster.